+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 -   TOTP support is reintroduced by popular demand. HOTP continues to be unsupported and heavily discouraged.
 -   Initial support for detecting and filling OTP fields with Autofill
 -   Importing TOTP secrets using QR codes
+-   Navigate into newly created folders and scroll to newly created passwords
 
 ## [1.9.2] - 2020-06-30